# modify & delete : UserRowMapper, UserListRowMapper 중복제거
##  modify
- RowMapper : return type 변경 ( T -> List<T> )
- UserRowMapper : 변경된 return type 적용
- JdbcTemplate : rowMapper 변경에 따른 selectOne 기능 수정
- UserDao : 사용하는 mapper 정리

## delete
- UserListRowMapper : 중복. 삭제
