+++ src/main/java/next/dao/mapper/row/RowMapper.java
@@ -2,7 +2,8 @@ package next.dao.mapper.row;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.List;
 
 public interface RowMapper<T> {
-    public T mapResult(ResultSet rs) throws SQLException;
+    public List<T> mapResult(ResultSet rs) throws SQLException;
 }
